Added "Contacts" and "Home" pages
The home page contains news that is “divided” into two sections. The Main News section contains news that was randomly selected on the topic “Ukraine”; the Side News section contains “hot” news taken from Ukraine.

Those news on the left have active links in their titles, and those on the right are entirely links.

On the contacts page I posted information about the location of the editorial office and the contact telephone number (the location was taken randomly, as well as the telephone number). Additionally, place a map with the geolocation of this address.